+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 3.0.2.0
+
+### Fixes
+* Fixed default animation value to true
+* Fixed stopPropagation on click event
+
+## 3.0.1.0
+* Updated selection logic
+* Added report page tooltip
+* Updated dependencies
+
+## 3.0.0.0
+* Updated powerbi-visuals-api to 5.10.0
+* Updated dependencies and fixed vulnerabilities
+* Added keyboard navigation and selection behavior
+* Migrated to formatting model
+* Added new settings to formatting pane
+
 ## 2.0.2
 * UPD: powerbi-visuals-tools has been updated to 3.0.9 to add IE11 support
 * UPD: API has been updated to 2.3.0 to add IE11 support